Further extract to common
