+++ nomos-services/data-availability/network/src/backends/libp2p/common.rs
@@ -7,11 +7,32 @@ use nomos_da_network_core::protocols::sampling;
 use nomos_da_network_core::protocols::sampling::behaviour::{
     BehaviourSampleReq, BehaviourSampleRes, SamplingError,
 };
-use nomos_da_network_core::swarm::validator::ValidatorEventsStream;
+use nomos_da_network_core::swarm::validator::{ValidatorEventsStream, ValidatorSwarm};
 use nomos_da_network_core::SubnetworkId;
+use nomos_libp2p::secret_key_serde;
+use nomos_libp2p::{ed25519, Multiaddr, PeerId};
+use overwatch_rs::overwatch::handle::OverwatchHandle;
+use serde::{Deserialize, Serialize};
+use std::collections::HashSet;
+use std::fmt::Debug;
+use subnetworks_assignations::MembershipHandler;
 use tokio::sync::mpsc::error::SendError;
 use tokio::sync::mpsc::UnboundedSender;
 use tokio::sync::{broadcast, mpsc};
+use tokio::task::JoinHandle;
+
+const BROADCAST_CHANNEL_SIZE: usize = 128;
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct DaNetworkBackendSettings<Membership> {
+    // Identification Secp256k1 private key in Hex format (`0x123...abc`). Default random.
+    #[serde(with = "secret_key_serde", default = "ed25519::SecretKey::generate")]
+    pub node_key: ed25519::SecretKey,
+    /// Membership of DA network PoV set
+    pub membership: Membership,
+    pub addresses: Vec<(PeerId, Multiaddr)>,
+    pub listening_address: Multiaddr,
+}
 
 /// Sampling events coming from da network
 #[derive(Debug, Clone)]
@@ -26,6 +47,91 @@ pub enum SamplingEvent {
     },
     /// A failed sampling error
     SamplingError { error: SamplingError },
+}
+
+pub(crate) struct InitializedBackendData {
+    pub task: JoinHandle<()>,
+    pub replies_task: JoinHandle<()>,
+    pub sampling_request_channel: UnboundedSender<(SubnetworkId, BlobId)>,
+    pub sampling_broadcast_receiver: broadcast::Receiver<SamplingEvent>,
+    pub verifying_broadcast_receiver: broadcast::Receiver<DaBlob>,
+}
+
+pub(crate) fn initialize_backend<Membership>(
+    config: DaNetworkBackendSettings<Membership>,
+    overwatch_handle: OverwatchHandle,
+) -> InitializedBackendData
+where
+    Membership: MembershipHandler<NetworkId = SubnetworkId, Id = PeerId>
+        + Clone
+        + Debug
+        + Send
+        + Sync
+        + 'static,
+{
+    let keypair = libp2p::identity::Keypair::from(ed25519::Keypair::from(config.node_key.clone()));
+    let (mut validator_swarm, events_streams) = ValidatorSwarm::new(
+        keypair,
+        config.membership.clone(),
+        config.addresses.clone().into_iter().collect(),
+    );
+    let sampling_request_channel = validator_swarm
+        .protocol_swarm()
+        .behaviour()
+        .sampling_behaviour()
+        .sample_request_channel();
+    let address = config.listening_address;
+    // put swarm to listen at the specified configuration address
+    validator_swarm
+        .protocol_swarm_mut()
+        .listen_on(address.clone())
+        .unwrap_or_else(|e| panic!("Error listening on DA network with address {address}: {e}"));
+
+    // Dial peers in the same subnetworks (Node might participate in multiple).
+    let local_peer_id = *validator_swarm.local_peer_id();
+    let mut connected_peers = HashSet::new();
+
+    config
+        .membership
+        .membership(&local_peer_id)
+        .iter()
+        .flat_map(|subnet| config.membership.members_of(subnet))
+        .filter(|peer| peer != &local_peer_id)
+        .filter_map(|peer| {
+            config
+                .addresses
+                .iter()
+                .find(|(p, _)| p == &peer)
+                .map(|(_, addr)| (peer, addr.clone()))
+        })
+        .for_each(|(peer, addr)| {
+            // Only dial if we haven't already connected to this peer.
+            if connected_peers.insert(peer) {
+                validator_swarm
+                    .dial(addr)
+                    .expect("Node should be able to dial peer in a subnet");
+            }
+        });
+
+    let task = overwatch_handle.runtime().spawn(validator_swarm.run());
+    let (sampling_broadcast_sender, sampling_broadcast_receiver) =
+        broadcast::channel(BROADCAST_CHANNEL_SIZE);
+    let (verifying_broadcast_sender, verifying_broadcast_receiver) =
+        broadcast::channel(BROADCAST_CHANNEL_SIZE);
+    let replies_task = overwatch_handle
+        .runtime()
+        .spawn(handle_validator_events_stream(
+            events_streams,
+            sampling_broadcast_sender,
+            verifying_broadcast_sender,
+        ));
+    InitializedBackendData {
+        task,
+        replies_task,
+        sampling_request_channel,
+        sampling_broadcast_receiver,
+        verifying_broadcast_receiver,
+    }
 }
 
 /// Task that handles forwarding of events to the subscriptions channels/stream
@@ -98,7 +204,7 @@ pub(crate) async fn handle_validator_events_stream(
     }
 }
 
-pub async fn handle_sample_request(
+pub(crate) async fn handle_sample_request(
     sampling_request_channel: &UnboundedSender<(SubnetworkId, BlobId)>,
     subnetwork_id: SubnetworkId,
     blob_id: BlobId,
